Add package for iPBot
