+++ src/main/java/DukePackage/DukeException.java
@@ -1,0 +1,7 @@
+package DukePackage;
+
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
