+++ src/main/java/DukePackage/TaskType.java
@@ -1,0 +1,7 @@
+package DukePackage;
+
+public enum TaskType {
+    TODO,
+    DEADLINE,
+    EVENT
+}
